backend/requirements.txt
fastapi
uvicorn
prisma
passlib[bcrypt]
PyJWT
python-dotenv